requirements.txt
streamlit
scikit-learn
tensorflow
joblib
numpy
pandas
matplotlib
Pillow
h5py
